+++ core/dbt/deprecations.py
@@ -82,15 +82,6 @@ class ModelsKeyNonModelDeprecation(DBTDeprecation):
     '''
 
 
-class AdapterMacroDeprecation(DBTDeprecation):
-    _name = 'adapter-macro'
-    _description = '''\
-    The "adapter_macro" macro has been deprecated. Instead, use the
-    `adapter.dispatch` method to find a macro and call the result.
-    adapter_macro was called for: {macro_name}
-    '''
-
-
 class PackageRedirectDeprecation(DBTDeprecation):
     _name = 'package-redirect'
     _description = '''\
@@ -140,7 +131,6 @@ deprecations_list: List[DBTDeprecation] = [
     DispatchPackagesDeprecation(),
     NotADictionaryDeprecation(),
     ModelsKeyNonModelDeprecation(),
-    AdapterMacroDeprecation(),
     PackageRedirectDeprecation()
 ]